WIP - Merging Mainchain and Sidechain endpoints
